Move SamplesTable away from Experiment component
