+++ src/containers/SamplesTable/MetadataAnnotationsCell.js
