+++ src/containers/SamplesTable/MetadataAnnotationsCell.scss
