+++ src/containers/SamplesTable/ProcessingInformation/ProcessingInformationModalContent.js
@@ -110,7 +110,8 @@ export default class ProcessingInformationModalContent extends React.Component {
             className="link"
           >
             our documentation
-          </a>.
+          </a>
+          .
         </div>
 
         <ProcessorVersion {...this.props} />
@@ -156,16 +157,19 @@ function AffymetrixScanProtocol() {
           className="link"
         >
           DOI: 10.1016/j.ygeno.2012.08.003
-        </a>){' '}
+        </a>
+        ){' '}
         <span>
-          and the SCAN.UPC bioconductor package documentation (<a
+          and the SCAN.UPC bioconductor package documentation (
+          <a
             href="https://doi.org/10.18129/B9.bioc.SCAN.UPC"
             target="_blank"
             rel="noopener noreferrer"
             className="link"
           >
             DOI: 10.18129/B9.bioc.SCAN.UPC
-          </a>).
+          </a>
+          ).
         </span>
       </p>
     </div>
@@ -191,7 +195,8 @@ function IlluminaScanProtocol() {
           className="link"
         >
           DOI: 10.1016/j.ygeno.2012.08.003
-        </a>).
+        </a>
+        ).
       </p>
     </div>
   );
@@ -245,9 +250,10 @@ function TxtimportProtocol() {
           className="link"
         >
           "lengthScaledTPM"
-        </a>, which are gene-level count-scale values that are generated by
-        scaling TPM using the average transcript length across samples and to
-        the library size. Note that tximport is applied at the{' '}
+        </a>
+        , which are gene-level count-scale values that are generated by scaling
+        TPM using the average transcript length across samples and to the
+        library size. Note that tximport is applied at the{' '}
         <em>experiment-level</em> rather than to single samples. For additional
         information, see the{' '}
         <a
@@ -257,7 +263,8 @@ function TxtimportProtocol() {
           className="link"
         >
           tximport Bioconductor page
-        </a>, the{' '}
+        </a>
+        , the{' '}
         <a
           href="http://bioconductor.org/packages/release/bioc/vignettes/tximport/inst/doc/tximport.html"
           target="_blank"
@@ -266,7 +273,8 @@ function TxtimportProtocol() {
         >
           tximport tutorial{' '}
           <em>Importing transcript abundance datasets with tximport</em>
-        </a>, and{' '}
+        </a>
+        , and{' '}
         <a
           href="http://dx.doi.org/10.12688/f1000research.7563.1"
           target="_blank"
