+++ src/containers/SamplesTable/ProcessingInformation/SubmitterSuppliedProtocol.scss
